Removed Stream End Time Menu Field, Set Stream Start Time to Date.now()
