+++ views/index.html
@@ -119,10 +119,6 @@
                                     <label class="ui-label">Name: </label>
                                     <input type="text" ng-model="name" name="name">
                                 </div>
-                                <div>
-                                    <label class="ui-label">End Date: </label>
-                                    <input type="text" ng-model="endDate" name="endDate">
-                                </div>
                                 <div class="full-width">
                                     <button class="inline ui-button" ng-click="createStream();toggleDropdown($event)">Add Stream</button>
                                     <button class="inline ui-button"  ng-click="toggleDropdown($event)">Cancel</button>
